+++ app/actions/homePageActions.js
@@ -1,6 +1,4 @@
-/**
- * @flow
- */
+// @flow
 export function setActiveMode(activeMode: string, activeModeOptions: Object = {}) {
   return {
     type: 'SET_ACTIVE_MODE',
