+++ app/api/torrents/TorrentAdapter.js
@@ -1,9 +1,5 @@
-/**
- * @flow
- * @param   {string} imdbId
- * @param   {object} extendedDetails
- */
 /* eslint global-require: 0 */
+// @flow
 import {
   determineQuality,
   formatSeasonEpisodeToString,
