+++ app/reducers/index.js
@@ -1,9 +1,11 @@
 import { combineReducers } from 'redux';
 import { routerReducer as routing } from 'react-router-redux';
 import homePageReducer from './homePageReducer';
+import itemPageReducer from './itemPageReducer';
 
 const rootReducer = combineReducers({
   homePageReducer,
+  itemPageReducer,
   routing
 });
